backend/requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Validation
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# 2FA
pyotp==2.9.0
qrcode[pil]==8.0

# Cache / Session
redis==5.2.1

# CORS
# (built into FastAPI)

# HD Wallet
hdwallets==0.1.2
mnemonic==0.21
eth-keys==0.6.1
eth-utils==5.2.0
base58==2.1.1

# Blockchain (BSC)
web3==6.21.0

# Utils
httpx==0.28.1
